In Flickr OAuth implementation, 'oauth_token_secret' key is returned after 'oauth/request_token' request. The key is needed for signing 'oauth/access_token' request later and this commit attempts to fix it.
The current AFOAuth1Client implementation only assign `accessToken` after 'oauth/access_token' request. This may work for Twitter, but this will not work for Flickr.
The signing method looks for `accessToken` when attempting to build the secret key needed, so we had to wrap the following logic by temporary assigning `accessToken` with `requestToken` to trick the signing method to sign with the correct key.
After making this change, I ran the sample code and it seems to work well for twitter as well.