Exclude "oauth_*" (aka "protocol parameters") from the body of POST requests.
The OAuth Specification says that requests should include the protocol parameters in EITHER the Authorization: HTTP header OR the request body, and to prefer the Header method. We're already using the header, so we can skip including them in the request body.

The relevant section of the specification is: http://tools.ietf.org/html/rfc5849#section-3.5.

(Hopefully my Objective-C is OK! Corrections welcomed! :smile:)